✨ Fallback to previously loaded translations
When the translations are retrieved, they are now stored forever in a dedicated cache key.
If, for whatever reasons, the newest translations cannot be retrieved, the previously loaded translations will be used.

This should mean that, in theory, even if the translations server is not responsive, not app should suffer from that.