⚡️ v1.1.0 - Improve speed in tests
To prevent hitting the translations API during unit/feature tests, it now stores the translations inside a JSON file once and re-use that file for the next tests. Note that it will never be refreshed by itself, you are required to delete the file yourself if you need the newest translations for some reason.

You can customize the file disk via the env variable TRUSTUP_IO_TRANSLATIONS_TESTS_STORAGE_DISK if you want. It defaults to "local".

I've improved the comments on the config file to improve DX.