✨ v1.3.0 - Disk cache support
Added support for a disk cache system, storing the translations locally in a .json file in the chosen disk and reusing it when needed.

This takes priority over the default cache system if both are enabled.

It makes it easier to see if the translations are already loaded locally and possibly to delete them.

It also makes more project compatible with the integration, as trustup-pro-api's cache is overriden by the tenancy package and doesn't run that well. This project can now define a custom disk and be fine.